Server test added
